// jobup-frontend/src/app/generated-sources/openapi/model/jobPostDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CommentDto } from './commentDto';


export interface JobPostDto { 
    id?: string;
    title?: string;
    description?: string;
    location?: string;
    createdAt?: string;
    createdById?: string;
    createdByName?: string;
    likes?: Array<string>;
    savedBy?: Array<string>;
    comments?: Array<CommentDto>;
    attachmentFileIds?: Array<string>;
}